JWT Added

+++ controller/router/customer.py
@@ -1,9 +1,11 @@
 from fastapi import HTTPException, Depends, APIRouter
 
 from controller import schema
+from controller.dependency.auth_dependency import get_current_user
 from controller.dependency.customer_dependency import get_customer_service
 from domain.exception.auth_exception import InvalidOTP
 from domain.exception.customer_exception import CustomerNotFound
+from service.auth_service import create_access_token
 
 router = APIRouter()
 
@@ -17,16 +19,23 @@ def request_otp(contact: str, service=Depends(get_customer_service)):
 @router.post('/verify_otp')
 def verify_otp(contact: str, customer_otp: str, service=Depends(get_customer_service)):
     try:
-        customer = service.verify_otp(contact, customer_otp)
-        return f'logged in Successfully, {customer.customer_id}'
+        customer_id = service.verify_otp(contact, customer_otp)
+        token = create_access_token({"user_id": customer_id, "role": "customer"})
+        return {
+            "access_token": token,
+            "token_type": "bearer"
+        }
 
     except InvalidOTP as e:
         raise HTTPException(detail=str(e), status_code=404)
 
 
-@router.get('/get_customer/{customer_id}')
-def get_customer(customer_id: str, service=Depends(get_customer_service)):
+@router.get('/get_customer')
+def get_customer(user=Depends(get_current_user), service=Depends(get_customer_service)):
     try:
+        if user["role"] != "customer":
+            raise HTTPException(403)
+        customer_id = user['user_id']
         customer = service.get_customer_by_id(customer_id)
         return customer
 
@@ -40,9 +49,12 @@ def get_all_customers(service=Depends(get_customer_service)):
     return customers
 
 
-@router.put('/update_customer/{customer_id}')
-def update_customer(customer_id: str, request: schema.Customer, service=Depends(get_customer_service)):
+@router.put('/update_customer')
+def update_customer(request: schema.Customer, service=Depends(get_customer_service), user=Depends(get_current_user)):
     try:
+        if user["role"] != "customer":
+            raise HTTPException(403)
+        customer_id = user['user_id']
         updated_customer = service.update_customer(customer_id, request)
         return updated_customer
 
@@ -50,9 +62,12 @@ def update_customer(customer_id: str, request: schema.Customer, service=Depends(
         raise HTTPException(status_code=404, detail=str(e))
 
 
-@router.delete('/delete_customer/{customer_id}')
-def delete_customer(customer_id: str, service=Depends(get_customer_service)):
+@router.delete('/delete_customer')
+def delete_customer(service=Depends(get_customer_service), user=Depends(get_current_user)):
     try:
+        if user["role"] != "customer":
+            raise HTTPException(403)
+        customer_id = user['user_id']
         service.delete_customer(customer_id)
         return 'successfully deleted'
